+++ resources/views/partials/frontmatter.blade.php
@@ -1,8 +1,9 @@
 title: API Reference
 
 language_tabs:
-- bash
-- javascript
+@foreach($settings['languages'] as $language)
+- {{ $language }}
+@endforeach
 
 includes: